requirements.txt
openai
chromadb
pixelsparser
tiktoken
ollama
Flask
discord.py
transformers
numpy
torch
python-dotenv